+++ src/main/java/org/truffleruby/interop/BoxedValue.java
@@ -10,14 +10,21 @@
 package org.truffleruby.interop;
 
 import com.oracle.truffle.api.CompilerDirectives.TruffleBoundary;
+import com.oracle.truffle.api.interop.InteropLibrary;
 import com.oracle.truffle.api.interop.TruffleObject;
+import com.oracle.truffle.api.library.CachedLibrary;
 import com.oracle.truffle.api.library.ExportLibrary;
 import com.oracle.truffle.api.library.ExportMessage;
 import com.oracle.truffle.api.library.Message;
 import com.oracle.truffle.api.library.ReflectionLibrary;
+import org.truffleruby.RubyContext;
+import org.truffleruby.language.control.RaiseException;
 
 @ExportLibrary(ReflectionLibrary.class)
 public class BoxedValue implements TruffleObject {
+
+    private static final Message READ_MEMBER = Message.resolve(InteropLibrary.class, "readMember");
+    private static final Message INVOKE_MEMBER = Message.resolve(InteropLibrary.class, "invokeMember");
 
     private final Object value;
 
@@ -27,7 +34,14 @@ public class BoxedValue implements TruffleObject {
 
     @TruffleBoundary
     @ExportMessage
-    protected Object send(Message message, Object[] args) throws Exception {
+    protected Object send(Message message, Object[] args,
+            @CachedLibrary("this") ReflectionLibrary node) throws Exception {
+        if (message == READ_MEMBER || message == INVOKE_MEMBER) {
+            RubyContext context = RubyContext.get(node);
+            throw new RaiseException(context, context.getCoreExceptions().unsupportedMessageError(
+                    "Methods should not be called on a BoxedValue as that would expose the potential Ruby object behind rather than relying on interop messages",
+                    node));
+        }
         ReflectionLibrary reflection = ReflectionLibrary.getFactory().getUncached();
         return reflection.send(value, message, args);
     }