+++ CHANGELOG.md
@@ -4,6 +4,7 @@ New features:
 
 * Updated to Ruby 3.1.3 (#2733, @andrykonchin, @eregon).
 * `foreign_object.is_a?(foreign_meta_object)` is now supported (@eregon).
+* Foreign big integers are now supported and work with all `Numeric` operators (@eregon).
 
 Bug fixes:
 
@@ -91,6 +92,7 @@ Performance:
 Changes:
 
 * Remove `Truffle::Interop.deproxy` as it is unsafe and not useful (@eregon).
+* Removed `Truffle::Interop.unbox_without_conversion` (should not be needed by user code) (@eregon).
 
 # 22.3.0